add : Book 관련 클래스들 > 인터페이스와 제어역전Ioc 적용, (스프링빈등록O, dbms연결0, 메모리에저장O)

+++ src/main/java/com/group/libraryapp/domain/Book.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @AllArgsConstructor
 public class Book {
     final private String name;
-    final private Integer price;
+    final private Long price;
 }
