+++ src/main/java/com/group/libraryapp/service/BookService.java
@@ -1,6 +1,6 @@
 package com.group.libraryapp.service;
 
-import antlr.StringUtils;
+import com.group.libraryapp.repository.book.BookMySqlRepository;
 import com.group.libraryapp.repository.book.BookRepository;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +12,7 @@ public class BookService {
         this.bookRepository = bookRepository;
     }
 
-    public void saveBook(String name, Integer price){
+    public void saveBook(String name, Long price){
         // TODO: 특수 문자열(!,$,%,')에 대한 유효성 검사가 필요함.
         if(name.isEmpty() && price != null) {
             throw new IllegalArgumentException();
